+++ ChangeLog.rst
@@ -7,6 +7,12 @@
 
 * Now disabled project will be checked by cron too. Except those project
   which were disabled manually by user.
+* Added a helper function ``find-projects-with-conflicting-systems``
+  which can be called manually after the import from the Quicklisp to
+  detect conflicts in system's names.
+
+  We'll need to automate this in regular checks and disable projects
+  which introduce conflicts.
 
 0.4.2 (2019-03-19)
 ==================